Fix: ProfileTap components로 분리
