+++ najuha-fronted/src/pages/Profilepage.js
@@ -2,6 +2,7 @@ import React from 'react'
 import Navbar from '../components/Navbar'
 import Profilesection from '../components/Profilesection'
 import Footer from '../components/Footer'
+
 function Profilepage() {
   return (
     <>
